libtracker-miner: Reduce TrackerDecorator file cache size
Instead of keeping all items in memory, a small cache is kept of upcoming
files to be processed, querying the database again for files missing the
nie:dataSource when the cache has been flushed. When no further items are
processed, the miner is stopped. Whenever the prioritized rdf:types list
changes, the cache is dropped, just to query again with the new order kept
into account.

This reduces the potential memory usage of tracker-extract, it would
previously would grow linearly with the number of files left to be
inspected, which might bring in some memory pressure on certain scenarios.

On the downside, this obviously makes the decorator a little bit slower,
due to more queries happening (as opposed to a huge one), and at times the
decorator is idle waiting to know what else to do.

Conflicts:
	tracker/src/libtracker-miner/tracker-decorator-internal.h
	tracker/src/libtracker-miner/tracker-decorator.c