[miner] Fix mining of files whose data was inserted by app
If an app inserts file data into the store before a file is created,
the tracker-miner will misbehave and construct bad sparql statements
instead of extracting more data from the new file.

User-visible effect: new pictures don't show up in Gallery, etc.

This patch to fix it was taken from
  https://bugzilla.gnome.org/show_bug.cgi?id=678986

Upstream is still debating its inclusion because of performance
concerns, but at least it fixes the problem for us.