[tracker] Update to 3.1.2. Contributes to JB#53910 OMP#JOLLA-60
Biggest difference to version 2 is that there's no more daemon on the
main package, only the one in tracker-miners. No more configuration
here either.

Speaking of the configuration, the tracker-configs.sh
seemed to be already conflicting with similar configuration
override in tracker-miners. Miners is, and was, overriding the
schema while this executed setting override as user.
Wondering if devices thus have wrong effective settings.