Relax the indexing rules for android_storage.
Currently Android Apps store their photos in their own direcotories and
as we do not index those these files do not ever endup to our Gallery.
Eventough this rules is done, some files are still not ending up to our
gallery as many apps also include .nomedia to their photo directories.

Because of above this does not change much so we might as well relax
this.

[config] Relax android_storage indexing. Contributes to JB#54611

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>