[tracker] Improve sorting by timestamp. JB#55106 OMP#JOLLA-289
Date time values were conditionally stored as either seconds from unix
epoch integers or ISO 8601 strings depending on whether the source date
time included microseconds or time zone offset data that could not be
encoded in an integer.

To resolve these types not being directly comparable a transform
function SparqlTimestamp is applied to the index created for date time
columns. But sorting query results using a pre-constructed index is an
optimization that is selectively utilized by sqlite because it is only
an optimization sometimes. And for queries where it is not an
optimization the result set will fall back to sorting on the column data
directly, which means comparing integers to strings with no context and
the integer values mostly end up sorting after the strings (everything's
a string in sqlite still).

Always storing the timestamp as an ISO 8601 string is the smallest
change that will produce mainly sane results. Date times with time zone
offsets will sort by local time, which is (probably) not desirable but
its an improvement on completely partioning with time zone values from
without. A logical next to resolve that issue would be to specify a
collation function that converted to utc which could then be installed
on the column itself, and not the just the indepedent index. That's a
schema change though.